requirements.txt
Mastodon.py==1.8.1
tracery==0.1.1
